Update camera.cpp
I've added the camera files, however I was using my custom matrix and vector classes with it before. I've got it switched over to glm, but I haven't compiled it just yet (so there's no guarantee it'll even compile right now, let alone work correctly). I'll fix it as soon as we have something to test this with