Libraries and camera
I've included all of the libraries and include files we'll need into
this file. It might be a little bloated, but I think that it'll be able
to work as is, and you won't have to do any additional setup to be able
to work on the solution.

Also fixed compiler errors in the camera code